Reorganize e2e tests
Move browser e2e tests to browser/ and shared utils to shared/
