+++ web/package.json
@@ -6,10 +6,9 @@
   "license": "MIT",
   "scripts": {
     "test": "jest --testPathIgnorePatterns e2e",
-    "test-e2e": "jest --coverage=false --runInBand index.e2e chrome.e2e",
-    "test-regression": "jest --coverage=false --runInBand regression",
-    "test-phabricator-e2e": "jest --coverage=false --runInBand phabricator.e2e",
-    "test-e2e-sgdev": "env SOURCEGRAPH_BASE_URL=https://sourcegraph.sgdev.org OVERRIDE_AUTH_SECRET=${SGDEV_OVERRIDE_AUTH_SECRET} jest --coverage=false e2e",
+    "test-e2e": "jest --coverage=false --runInBand ./e2e/e2e.test",
+    "test-regression": "jest --coverage=false --runInBand ./e2e/regression.test",
+    "test-e2e-sgdev": "env SOURCEGRAPH_BASE_URL=https://sourcegraph.sgdev.org OVERRIDE_AUTH_SECRET=${SGDEV_OVERRIDE_AUTH_SECRET} jest --coverage=false --runInBand ./e2e/e2e.test",
     "build": "cross-env NODE_OPTIONS=\"--max_old_space_size=4096\" gulp build",
     "watch": "cross-env NODE_OPTIONS=\"--max_old_space_size=4096\" gulp watch",
     "watch-webpack": "cross-env NODE_OPTIONS=\"--max_old_space_size=4096\" gulp watchWebpack",
