+++ web/src/e2e/e2e.test.ts
@@ -1,7 +1,7 @@
 import * as path from 'path'
-import { saveScreenshotsUponFailuresAndClosePage } from '../../../shared/src/util/screenshotReporter'
-import { retry } from '../util/e2e-test-utils'
-import { baseURL, createDriverForTest, Driver, gitHubToken, percySnapshot } from './util'
+import { saveScreenshotsUponFailuresAndClosePage } from '../../../shared/src/e2e/screenshotReporter'
+import { retry } from '../../../shared/src/e2e/e2e-test-utils'
+import { baseURL, createDriverForTest, Driver, gitHubToken, percySnapshot } from '../../../shared/src/e2e/driver'
 import got from 'got'
 import { gql } from '../../../shared/src/graphql/graphql'
 import { random } from 'lodash'
